🛠️ refactor: (BE) - 소켓 관련 상수 수정
