+++ backend/socket/src/constant/event.constant.ts
@@ -1,0 +1,12 @@
+export enum ROOM_EVENT {
+	CREATE_ROOM = 'create_room',
+	ENTER_ROOM = 'enter_room',
+	USER_ENTER = 'user_enter',
+	LEAVE_ROOM = 'leave_room',
+	START_INTERVIEW = 'start_interview',
+	JOIN_INTERVIEW = 'join_interview',
+	END_INTERVIEW = 'end_interview',
+	END_FEEDBACK = 'end_feedback',
+	COUNT_FEEDBACK = 'count_feedback',
+	TERMINATE_SESSION = 'terminate_session',
+}
