+++ backend/socket/src/constant/index.ts
@@ -1,1 +1,3 @@
 export * from './room.constant';
+export * from './user.constant';
+export * from './event.constant';
