+++ backend/socket/src/constant/user.constant.ts
@@ -1,0 +1,5 @@
+export enum USER_ROLE {
+	INETRVIEWER = 'interviewer',
+	INTERVIEWEE = 'interviewee',
+	NONE = 'none',
+}
